src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
addict=2.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
aom=3.6.0=hd77b12b_0
asttokens=3.0.0=pyhd8ed1ab_1
black=25.1.0=pypi_0
blas=1.0=mkl
brotlicffi=1.0.9.2=py310h5da7b33_1
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.8.3=h4c7d964_0
cairo=1.16.0=h85cdd14_6
certifi=2025.8.3=pyhd8ed1ab_0
cffi=1.17.1=py310h827c3e9_1
cfgv=3.4.0=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.1=pypi_0
colorama=0.4.6=py310haa95532_0
comm=0.2.3=pyhe01879c_0
contourpy=1.3.2=pypi_0
cpython=3.10.18=py310hd8ed1ab_0
cycler=0.12.1=pypi_0
dav1d=1.2.1=h2bbff1b_0
debugpy=1.8.15=py310h699e580_0
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=py310haa95532_0
distlib=0.4.0=pypi_0
eigen=3.4.0=h59b6b97_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.1=h8ddb27b_0
ffmpeg=6.1.1=hc79a5da_3
filelock=3.17.0=py310haa95532_0
fire=0.7.0=pypi_0
fontconfig=2.14.1=hb33846d_3
fonttools=4.59.0=pypi_0
freeglut=3.4.0=h8a1e904_1
freetype=2.13.3=h0620614_0
fribidi=1.0.10=h62dcd97_0
fsspec=2025.7.0=py310hbc747e5_0
gflags=2.2.2=hd77b12b_1
giflib=5.2.2=h7edc060_0
glog=0.5.0=hd77b12b_1
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py310h827c3e9_0
graphite2=1.3.14=hd77b12b_1
groundingdino=0.1.0=dev_0
gst-plugins-base=1.24.12=h91a6125_1
gstreamer=1.24.12=hfb93a4f_1
gstreamer-orc=0.4.41=ha00e802_0
harfbuzz=10.2.0=he2f9f60_1
hdf5=1.14.5=ha36df97_2
huggingface_hub=0.31.2=py310haa95532_0
hydra-core=1.3.2=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
identify=2.6.12=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
iopath=0.1.10=pypi_0
ipykernel=6.30.1=pyh3521513_0
ipython=8.37.0=pyha7b4d00_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=py310haa95532_0
jpeg=9e=h827c3e9_3
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh5737063_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=he4457a5_1
lcms2=2.16=h62be587_1
lerc=4.0.0=h5da7b33_0
libabseil=20250127.0=cxx17_h52369b4_0
libavif=1.1.1=h827c3e9_0
libdeflate=1.22=h5bf469e_0
libffi=3.4.4=hd77b12b_1
libglib=2.84.2=h405b238_0
libiconv=1.16=h2bbff1b_3
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=17.4=h4a159e6_1
libprotobuf=5.29.3=h65a231f_1
libsodium=1.0.20=hc70643c_0
libtheora=1.1.1=h8cc25b3_3
libtiff=4.7.0=h404307b_0
libtorch=2.5.1=cpu_mkl_h9002858_104
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h3d04722_1
libxml2=2.13.8=h866ff63_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=3.0.2=py310h827c3e9_0
matplotlib=3.10.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h827c3e9_2
mkl_fft=1.3.11=py310h827c3e9_0
mkl_random=1.2.8=py310hc64d2fc_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py310haa95532_0
mypy-extensions=1.1.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=py310haa95532_0
nodeenv=1.9.1=pypi_0
numpy=2.0.1=py310h055cbcc_1
numpy-base=2.0.1=py310h65a83cf_1
omegaconf=2.3.0=pypi_0
opencv=4.10.0=py310h28596fa_7
openjpeg=2.5.2=h9b5d1b5_1
openssl=3.5.1=h725018a_0
opentelemetry-api=1.30.0=py310haa95532_0
packaging=25.0=py310haa95532_0
pandas=2.3.1=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pathspec=0.12.1=pypi_0
pcre2=10.42=h0ff8eda_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=py310hb328d1f_0
pip=25.2=pypi_0
pixman=0.40.0=h2bbff1b_1
platformdirs=4.3.8=pyhe01879c_0
portalocker=3.2.0=pypi_0
pre-commit=4.2.0=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
psutil=7.0.0=py310ha8f682b_0
pure_eval=0.2.3=pyhd8ed1ab_1
py-cpuinfo=9.0.0=pypi_0
pybboxes=0.1.6=pypi_0
pycocotools=2.0.10=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.19.2=pyhd8ed1ab_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.18=h981015d_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python_abi=3.10=2_cp310
pytz=2025.2=pypi_0
pywin32=311=py310h282bd7d_0
pyyaml=6.0.2=py310h827c3e9_0
pyzmq=27.0.1=py310h49f0f51_0
qtbase=6.7.3=hd088775_3
regex=2024.11.6=py310h827c3e9_0
requests=2.32.4=py310haa95532_0
ruff=0.12.7=pypi_0
safetensors=0.5.3=py310haa69e8e_0
sahi=0.11.31=pypi_0
sam-2=1.0=pypi_0
scipy=1.15.3=pypi_0
segment-anything=1.0=pypi_0
setuptools=78.1.1=py310haa95532_0
shapely=2.1.1=pypi_0
six=1.17.0=pyhe01879c_1
sleef=3.5.1=h8cc25b3_2
sqlite=3.50.2=hda9a48d_1
stack_data=0.6.3=pyhd8ed1ab_1
supervision=0.26.1=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
termcolor=3.1.0=pypi_0
terminaltables=3.1.10=pypi_0
timm=1.0.19=pypi_0
tk=8.6.14=h5e9d12e_1
tokenizers=0.15.1=py310h49fca51_0
tomli=2.2.1=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchvision=0.20.1+cu121=pypi_0
tornado=6.5.1=py310ha8f682b_0
tqdm=4.67.1=py310h9909e9c_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.39.3=py310haa95532_0
typing-extensions=4.12.2=py310haa95532_0
typing_extensions=4.12.2=py310haa95532_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
ultralytics=8.3.174=pypi_0
ultralytics-thop=2.0.15=pypi_0
urllib3=2.5.0=py310haa95532_0
vc=14.42=haa95532_5
vc14_runtime=14.44.35208=h4927774_10
virtualenv=20.32.0=pypi_0
vs2015_runtime=14.44.35208=ha6b5a95_10
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wrapt=1.17.0=py310h827c3e9_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
yapf=0.43.0=pypi_0
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=py310haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
